ENH: gramex.cache.open(excel) supports hyperlinks. Fixes #722
